# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Refyne API",
    "version": "1.0.0",
    "description": "LLM-powered web extraction API that transforms unstructured websites into clean, typed JSON."
  },
  "servers": [
    {
      "url": "https://api.refyne.uk",
      "description": "Production API"
    },
    {
      "url": "http://localhost:8080",
      "description": "Local Development"
    }
  ],
...
  "paths": {
    "/api/v1/extract": {
      "post": {
        "operationId": "extract",
        "tags": ["Extraction"],
        "summary": "Extract data from URL",
        "description": "Extract structured data from a single webpage using an LLM. Define a schema specifying the data structure you want, and the API returns JSON matching that schema.",
        "security": [{"Authorization": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtractRequest"
              },
              "examples": {
                "basic": {
                  "summary": "Basic extraction",
                  "value": {
                    "url": "https://example.com/product/123",
                    "schema": {
                      "name": "string",
                      "price": "number",
                      "description": "string"
                    }
                  }
                },
                "with_llm_config": {
                  "summary": "With custom LLM provider",
                  "value": {
                    "url": "https://example.com/product/123",
                    "schema": {
                      "name": "string",
                      "price": "number"
                    },
                    "llm_config": {
                      "provider": "anthropic",
                      "model": "claude-3-5-sonnet-20241022"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful extraction",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtractResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "429": {
            "description": "Rate limit exceeded",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/crawl": {
      "post": {
        "operationId": "createCrawlJob",
        "tags": ["Crawling"],
        "summary": "Start crawl job",
        "description": "Start an asynchronous crawl job that extracts data from multiple pages. The crawler follows links matching your selectors or patterns and extracts data from each page.",
        "security": [{"Authorization": []}],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CrawlRequest"
              },
              "examples": {
                "with_selector": {
                  "summary": "Follow links by CSS selector",
                  "value": {
                    "url": "https://example.com/products",
                    "schema": {
                      "name": "string",
                      "price": "number"
                    },
                    "options": {
                      "follow_selector": "a.product-link",
                      "max_pages": 20
                    }
                  }
                },
                "with_pattern": {
                  "summary": "Follow links by URL pattern",
                  "value": {
                    "url": "https://example.com/products",
                    "schema": {
                      "name": "string",
                      "price": "number"
                    },
                    "options": {
                      "follow_pattern": "/products/[0-9]+",
                      "max_pages": 20
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Job created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CrawlJobCreated"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs": {
      "get": {
        "operationId": "listJobs",
        "tags": ["Jobs"],
        "summary": "List jobs",
        "description": "List all crawl jobs for the authenticated user, ordered by creation date descending.",
        "security": [{"Authorization": []}],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Number of jobs to return"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Offset for pagination"
          }
        ],
        "responses": {
          "200": {
            "description": "List of jobs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobList"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{id}": {
      "get": {
        "operationId": "getJob",
        "tags": ["Jobs"],
        "summary": "Get job status",
        "description": "Get the current status and details of a crawl job.",
        "security": [{"Authorization": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Job details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/jobs/{id}/results": {
      "get": {
        "operationId": "getJobResults",
        "tags": ["Jobs"],
        "summary": "Get job results",
        "description": "Get the extraction results from a completed crawl job. Use `merge=true` to get all results merged into a single object.",
        "security": [{"Authorization": []}],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Job ID"
          },
          {
            "name": "merge",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            },
            "description": "Merge all results into a single object with deduplicated arrays"
          }
        ],
        "responses": {
          "200": {
            "description": "Job results",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobResults"
                }
              }
            }
          },
          "404": {
            "description": "Job not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "Authorization": {
        "type": "http",
        "scheme": "bearer",
        "description": "Include your token in the Authorization header."
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "status": {
            "type": "integer"
          }
        }
      },
      "ExtractRequest": {
        "type": "object",
        "required": ["url", "schema"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "URL to extract data from"
          },
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "Schema defining the data structure to extract (JSON or YAML)"
          },
          "fetch_mode": {
            "type": "string",
            "enum": ["auto", "static", "dynamic"],
            "default": "auto",
            "description": "Fetch mode: auto (detect), static (fast), or dynamic (JavaScript rendering)"
          },
          "llm_config": {
            "$ref": "#/components/schemas/LlmConfig"
          }
        }
      },
      "ExtractResponse": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "additionalProperties": true,
            "description": "Extracted data matching the schema"
          },
          "url": {
            "type": "string",
            "description": "URL that was extracted"
          },
          "fetched_at": {
            "type": "string",
            "format": "date-time",
            "description": "Timestamp when the page was fetched"
          },
          "usage": {
            "$ref": "#/components/schemas/TokenUsage"
          },
          "metadata": {
            "$ref": "#/components/schemas/ExtractionMetadata"
          }
        }
      },
...
      "CrawlRequest": {
        "type": "object",
        "required": ["url", "schema"],
        "properties": {
          "url": {
            "type": "string",
            "format": "uri",
            "description": "Seed URL to start crawling from"
          },
          "schema": {
            "type": "object",
            "additionalProperties": true,
            "description": "Schema defining the data structure to extract"
          },
          "options": {
            "$ref": "#/components/schemas/CrawlOptions"
          },
          "webhook_url": {
            "type": "string",
            "format": "uri",
            "description": "URL to receive webhook on completion"
          },
          "llm_config": {
            "$ref": "#/components/schemas/LlmConfig"
          }
        }
      },
...
      "CrawlJobCreated": {
        "type": "object",
        "properties": {
          "job_id": {
            "type": "string",
            "description": "Unique job identifier"
          },
          "status": {
            "type": "string",
            "description": "Initial job status"
          },
          "status_url": {
            "type": "string",
            "description": "URL to check job status"
          }
        }
      },
...
    }
  }
}